Rename product_name/product_Name to product_name/package_product_name
