+++ src/product_name/package_product_name/dotnet.ts
